+++ src/main/java/baseball/baseballGame/BaseballPrinter.java
@@ -1,4 +1,4 @@
-package baseball.numGame.baseballGame;
+package baseball.baseballGame;
 
 import baseball.game.Printer;
