+++ src/main/java/baseball/game/RandomNumsGenerator.java
@@ -1,4 +1,4 @@
-package baseball.numGame;
+package baseball.game;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
